[#84] Refactor: Move OpenViduGenerateTokenRequest to request package and update validation in OpenViduWebhookEventRequest
