+++ src/main/java/soon/capstone/infrastructure/openvidu/controller/dto/request/OpenViduWebhookEventRequest.java
@@ -1,6 +1,8 @@
 package soon.capstone.infrastructure.openvidu.controller.dto.request;
 
+import jakarta.validation.constraints.Future;
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 import soon.capstone.infrastructure.openvidu.service.dto.request.OpenViduWebhookEventServiceRequest;
 
@@ -12,6 +14,8 @@ public record OpenViduWebhookEventRequest(
     @NotBlank(message = "body는 필수입니다.")
     String body,
 
+    @NotNull(message = "reservedAt은 필수입니다.")
+    @Future(message = "reservedAt은 현재 시간보다 미래여야 합니다.")
     LocalDateTime reservedAt
 
 ) {
